Revamping authentication test
